add llm config
