+++ src/llm/exec.rs
@@ -145,13 +145,14 @@ impl ExecutionPlan for AsyncFuncExec {
             // stream and satisfy lifetime requirements.
             let async_exprs_captured = Arc::clone(&async_exprs_captured);
             let schema_captured = schema_captured.clone();
+            let config_option = context.session_config().options().clone();
 
             async move {
                 let batch = batch?;
                 // append the result of evaluating the async expressions to the output
                 let mut output_arrays = batch.columns().to_vec();
                 for async_expr in async_exprs_captured.iter() {
-                    let output = async_expr.invoke_with_args(&batch).await?;
+                    let output = async_expr.invoke_with_args(&batch, &config_option).await?;
                     output_arrays.push(output.to_array(batch.num_rows())?);
                 }
                 let batch = RecordBatch::try_new(schema_captured, output_arrays)?;
